templates/django-base/requirements.txt
# Django Core
Django>=5.2,<6.0
psycopg2-binary>=2.9
dj-database-url>=2.1

# Django REST Framework & Auth
djangorestframework>=3.15
dj-rest-auth>=7.0
django-allauth>=65.0
django-authtools>=2.0

# Storage
django-storages>=1.14
boto3>=1.34

# GCP Deployment
django-environ>=0.11
google-auth>=2.27
google-cloud-secret-manager>=2.18
google-cloud-storage>=2.14
gunicorn>=21.2
whitenoise>=6.6
django-cors-headers>=4.3

# Task Queue
django-google-cloud-tasks>=2.20

# Development
python-dotenv>=1.0

# Deployment Automation
fabric>=3.2
